pwd and clu passes

+++ faiss/gpu/impl/TestComputeMode.cu
@@ -27,7 +27,7 @@ namespace faiss {
 		       Tensor<T, 4, true, int>& burst,
 		       Tensor<int, 5, true, int>& blocks,
 		       Tensor<T, 5, true, int>& centroids,
-		       Tensor<int, 4, true, int>& clusters,
+		       Tensor<uint8_t, 4, true, int>& clusters,
 		       Tensor<float, 1, true, int>& modes,
 		       int patchsize, float offset,
 		       cudaStream_t stream){
@@ -40,7 +40,7 @@ namespace faiss {
 		       Tensor<T, 4, true, int>& burst,
 		       Tensor<int, 5, true, int>& blocks,
 		       Tensor<T, 5, true, int>& centroids,
-		       Tensor<int, 4, true, int>& clusters,
+		       Tensor<uint8_t, 4, true, int>& clusters,
 		       Tensor<float, 1, true, int>& modes,
 		       int patchsize, float offset,
 		       cudaStream_t stream){
@@ -59,7 +59,7 @@ namespace faiss {
 			   Tensor<T, 4, true, int>& burst,
 			   Tensor<int, 5, true, int>& blocks,
 			   Tensor<T, 5, true, int>& centroids,
-			   Tensor<int, 4, true, int>& clusters,
+			   Tensor<uint8_t, 4, true, int>& clusters,
 			   Tensor<float, 1, true, int>& modes,
 			   int patchsize, float offset,
 			   cudaStream_t stream){
@@ -86,7 +86,7 @@ namespace faiss {
 			   Tensor<float, 4, true, int>& burst,
 			   Tensor<int, 5, true, int>& blocks,
 			   Tensor<float, 5, true, int>& centroids,
-			   Tensor<int, 4, true, int>& clusters,
+			   Tensor<uint8_t, 4, true, int>& clusters,
 			   Tensor<float, 1, true, int>& modes,
 			   int patchsize, float offset,
 			   cudaStream_t stream){
@@ -99,7 +99,7 @@ namespace faiss {
 			   Tensor<half, 4, true, int>& burst,
 			   Tensor<int, 5, true, int>& blocks,
 			   Tensor<half, 5, true, int>& centroids,
-			   Tensor<int, 4, true, int>& clusters,
+			   Tensor<uint8_t, 4, true, int>& clusters,
 			   Tensor<float, 1, true, int>& modes,
 			   int patchsize, float offset,
 			   cudaStream_t stream){
