+++ faiss/gpu/impl/TestComputeMode.cuh
@@ -16,7 +16,7 @@ namespace faiss {
 			   Tensor<float, 4, true, int>& burst,
 			   Tensor<int, 5, true, int>& blocks,
 			   Tensor<float, 5, true, int>& centroids,
-			   Tensor<int, 4, true, int>& clusters,
+			   Tensor<uint8_t, 4, true, int>& clusters,
 			   Tensor<float, 1, true, int>& modes,
 			   int patchsize, float offset,
 			   cudaStream_t stream);
@@ -26,7 +26,7 @@ namespace faiss {
 			   Tensor<half, 4, true, int>& burst,
 			   Tensor<int, 5, true, int>& blocks,
 			   Tensor<half, 5, true, int>& centroids,
-			   Tensor<int, 4, true, int>& clusters,
+			   Tensor<uint8_t, 4, true, int>& clusters,
 			   Tensor<float, 1, true, int>& modes,
 			   int patchsize, float offset,
 			   cudaStream_t stream);
