+++ faiss/gpu/impl/TestMeshSearchSpace.cu
@@ -26,7 +26,7 @@ namespace faiss {
 		       Tensor<T, 4, true, int>& burst,
 		       Tensor<int, 5, true, int>& blocks,
 		       Tensor<T, 5, true, int>& centroids,
-		       Tensor<int, 4, true, int>& clusters,
+		       Tensor<uint8_t, 4, true, int>& clusters,
 		       int patchsize, float offset,
 		       cudaStream_t stream){
 	int* one = (int*)malloc(sizeof(int));
@@ -51,7 +51,7 @@ namespace faiss {
 		       Tensor<T, 4, true, int>& burst,
 		       Tensor<int, 5, true, int>& blocks,
 		       Tensor<T, 5, true, int>& centroids,
-		       Tensor<int, 4, true, int>& clusters,
+		       Tensor<uint8_t, 4, true, int>& clusters,
 		       int patchsize, float offset,
 		       cudaStream_t stream){
 	fprintf(stdout,"test_case_1,\n");
@@ -69,7 +69,7 @@ namespace faiss {
 				Tensor<T, 4, true, int>& burst,
 				Tensor<int, 5, true, int>& blocks,
 				Tensor<T, 5, true, int>& centroids,
-				Tensor<int, 4, true, int>& clusters,
+				Tensor<uint8_t, 4, true, int>& clusters,
 				int patchsize, float offset,
 				cudaStream_t stream){
 
@@ -95,7 +95,7 @@ namespace faiss {
 				Tensor<float, 4, true, int>& burst,
 				Tensor<int, 5, true, int>& blocks,
 				Tensor<float, 5, true, int>& centroids,
-				Tensor<int, 4, true, int>& clusters,
+				Tensor<uint8_t, 4, true, int>& clusters,
 				int patchsize, float offset,
 				cudaStream_t stream){
       test_mesh_search_space<float>(test_case,dists,burst,blocks,centroids,clusters,
@@ -107,7 +107,7 @@ namespace faiss {
 				Tensor<half, 4, true, int>& burst,
 				Tensor<int, 5, true, int>& blocks,
 				Tensor<half, 5, true, int>& centroids,
-				Tensor<int, 4, true, int>& clusters,
+				Tensor<uint8_t, 4, true, int>& clusters,
 				int patchsize, float offset,
 				cudaStream_t stream){
       test_mesh_search_space<half>(test_case,dists,burst,blocks,centroids,clusters,
