+++ faiss/gpu/impl/TestMeshSearchSpace.cuh
@@ -16,7 +16,7 @@ namespace faiss {
 				Tensor<float, 4, true, int>& burst,
 				Tensor<int, 5, true, int>& blocks,
 				Tensor<float, 5, true, int>& centroids,
-				Tensor<int, 4, true, int>& clusters,
+				Tensor<uint8_t, 4, true, int>& clusters,
 				int patchsize, float offset,
 				cudaStream_t stream);
 
@@ -25,7 +25,7 @@ namespace faiss {
 				Tensor<half, 4, true, int>& burst,
 				Tensor<int, 5, true, int>& blocks,
 				Tensor<half, 5, true, int>& centroids,
-				Tensor<int, 4, true, int>& clusters,
+				Tensor<uint8_t, 4, true, int>& clusters,
 				int patchsize, float offset,
 				cudaStream_t stream);
